Move grid and axis code to a backgrounds folder
