+++ packages/perseus/src/widgets/interactive-graphs/backgrounds/axis-ticks.tsx
@@ -1,10 +1,10 @@
 import * as React from "react";
 
-import {useTransformVectorsToPixels} from "./graphs/use-transform";
-import {MAX, MIN, X, Y} from "./math";
-import useGraphConfig from "./reducer/use-graph-config";
+import {useTransformVectorsToPixels} from "../graphs/use-transform";
+import {MAX, MIN, X, Y} from "../math";
+import useGraphConfig from "../reducer/use-graph-config";
 
-import type {GraphDimensions} from "./types";
+import type {GraphDimensions} from "../types";
 import type {vec} from "mafs";
 
 const tickSize = 10;
