+++ packages/perseus/src/widgets/interactive-graphs/backgrounds/grid.tsx
@@ -1,12 +1,13 @@
 import {Coordinates} from "mafs";
 import * as React from "react";
 
+import {X, Y} from "../math";
+
 import AxisArrows from "./axis-arrows";
 import {AxisTicks} from "./axis-ticks";
-import {X, Y} from "./math";
 
-import type {GraphRange} from "../../perseus-types";
-import type {SizeClass} from "../../util/sizing-utils";
+import type {GraphRange} from "../../../perseus-types";
+import type {SizeClass} from "../../../util/sizing-utils";
 import type {vec} from "mafs";
 
 interface GridProps {
