+++ packages/perseus/src/widgets/interactive-graphs/backgrounds/legacy-grid.tsx
@@ -1,13 +1,12 @@
 import {View} from "@khanacademy/wonder-blocks-core";
 import * as React from "react";
 
-import AssetContext from "../../asset-context";
-import {SvgImage} from "../../components";
-import {interactiveSizes} from "../../styles/constants";
+import AssetContext from "../../../asset-context";
+import {SvgImage} from "../../../components";
+import {interactiveSizes} from "../../../styles/constants";
+import {X} from "../math";
 
-import {X} from "./math";
-
-import type {PerseusImageBackground} from "../../perseus-types";
+import type {PerseusImageBackground} from "../../../perseus-types";
 
 interface Props {
     box: [number, number];
